requirements.txt
# requirements.txt
google-cloud-vision
google-cloud-aiplatform
python-dotenv
networkx
matplotlib